pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[project]
name = "sat-finder"
version = "0.0.1"
description = "Find structures in satellite imagery using DINOv3 embeddings"
authors = [{ name = "Marco Willi", email = "mwilli-dev@proton.me" }]

dependencies = [
    # Core
    "numpy<2",  # Pin for PyTorch compatibility
    "pyyaml",

    # Deep Learning
    "torch",
    "torchvision",
    "timm",

    # Image processing
    "pillow>=10.0.0",

    # Gradio HF App
    "gradio>=4.0.0",
    "fastapi>=0.100",
    "uvicorn>=0.23",
]

[project.optional-dependencies]
dev = ["pytest", "ruff", "pre-commit", "python-dotenv", "pyprojroot", "jupyter", "ipykernel"]
